chore: split ViewerDialog into ViewerContent/TagsDrawer/ViewerMedia/PostNotFoundFallback, unify origin queryKey
